requirements.txt
crewai
crewai_tools
python-dotenv
rich
# Additional dependencies
snowflake-connector-python
